version info

+++ random_state.c
@@ -12,7 +12,7 @@ atomic_size_t COUNTER = 0;
 
 random_state_t new_state_from_keys(uint64_t *a, uint64_t *b) {
     ahasher_t hasher = hasher_from_random_state( a[0], a[1], a[2], a[3] );
-    uint64_t stack_position = (uint64_t)(&new_state_from_keys);
+    uint64_t stack_position = (ptrdiff_t)(&new_state_from_keys);
 #if defined(__arm__) || defined(__arm64__) || defined(__aarch64__) || defined(_M_ARM) || defined(_M_ARM64)
     uint64_t counter = (uint64_t)atomic_load_explicit(&COUNTER, memory_order_relaxed);
     counter += stack_position;
